BMP280.c: Fix variable types
Fix the type of variables for calibration. This
important patch has been provided by Ger Koper.

Suggested-by: Ger Koper <g.koper@xs4all.nl>

+++ BMP280/c/BMP280.c
@@ -35,18 +35,18 @@
 #define   BMP280_PRESSUREDATA 0xF7
 #define   BMP280_TEMPDATA     0xFA
 
-int cal_t1 = 27504;
-int cal_t2 = 26435;
-int cal_t3 = -1000;
-int cal_p1 = 36477;
-int cal_p2 = -10685;
-int cal_p3 = 3024;
-int cal_p4 = 2855;
-int cal_p5 = 140;
-int cal_p6 = -7;
-int cal_p7 = 15500;
-int cal_p8 = -14500;
-int cal_p9 = 6000;
+unsigned short int cal_t1 = 27504;
+short int cal_t2 = 26435;
+short int cal_t3 = -1000;
+unsigned short int cal_p1 = 36477;
+short int cal_p2 = -10685;
+short int cal_p3 = 3024;
+short int cal_p4 = 2855;
+short int cal_p5 = 140;
+short int cal_p6 = -7;
+short int cal_p7 = 15500;
+short int cal_p8 = -14500;
+short int cal_p9 = 6000;
 
 void load_calibration(int fd)
 {
